+++ lib/src/utils/constants.dart
@@ -1,3 +1,3 @@
 const kDefaultAdServerUrl = 'https://server.megabrain.co';
 const kSdkLabel = 'sdk-flutter';
-const kSdkVersion = '2.2.2';
+const kSdkVersion = '2.2.3-rc.2';